% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diverged_EC.R
\name{divergedEC}
\alias{divergedEC}
\title{Calculate Diverged Background Expression Conservation (EC) Distribution}
\usage{
divergedEC(
  expression = NULL,
  csM1 = NULL,
  csM2 = NULL,
  conv = 1e-04,
  maxIter = 100,
  threads = 1,
  weights_fixed = NULL
)
}
\arguments{
\item{expression}{The 'gene X condition' expression matrix for the first dataset.}

\item{csM1}{The 'gene X gene' correlation matrix for core genes from the first dataset.}

\item{csM2}{The 'gene X gene' correlation (sub)matrix from the second dataset. Gene IDs must correspond to those in the \code{expression} matrix.}

\item{conv}{Convergence criterion indicating the permissible difference in EC values between iterations for convergence.}

\item{maxIter}{The maximum number of iterations for EC calculation.}

\item{weights}{Optional: Weights for calculating weighted correlations. If not provided, the function iteratively refines weights in each iteration to calculate EC.}

\item{mc_cores}{Number of CPU cores to be used for parallel processing.}
}
\value{
A named vector of EC values representing a diverged or random distribution of gene expression conservation.
}
\description{
Computes the 'diverged' background EC distribution by calculating EC values between permutated gene expression data and a given correlation matrix. This function helps in assessing the conservation of gene expression patterns in a context where gene relationships are expected to have diverged.
}
\details{
The function permutes rows of the expression matrix and calculates the EC between each permutated row and the rows of the csM1 matrix. If weights are not provided, it employs an iterative approach to refine the weights in each iteration. This process aims to model the divergence in gene expression patterns.
}
\examples{
randomEC1 <- divergedEC(exprList$exprValues1, core_submatrix2_ordered,
                         ortho, EC$ECweights, conv = 0.001, maxIter = 200)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
