% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diverged_EC_fast.R
\name{divergedEC_fast}
\alias{divergedEC_fast}
\title{diverged EC distribution}
\usage{
divergedEC_fast(
  expression = exprList$exprValues1,
  csM1 = corM_ortho$csM1,
  csM2 = csM2_ordered,
  ortho,
  conv = 1e-04,
  maxIter = 100,
  threads = 1,
  weights = NA
)
}
\arguments{
\item{expression}{First 'gene X condition' expression matrix}

\item{csM1}{Correlation matrix of core genes - sample 1}

\item{csM2}{Correlation (sub)matrix. Gene IDs must be the same as used in the exprM}

\item{ortho}{Two-column dataframe of gene IDs in species1/sample1,
and the corresponding orthologous gene IDs in species2/sample2}

\item{conv}{Convergence criterion indicating how much the final EC value per gene can
differ from the result of the previous iteration}

\item{maxIter}{The maximum number of iterations}

\item{weights}{Optional argument: weights that will be used for calculating the weighted correlation.
If no weights are provided, the EC value will be calculated iteratively, refining the weights each iteration}

\item{mc_cores}{Number of cpus to be used}
}
\value{
A named vector of EC values representing a random distribution
}
\description{
Calculates the 'diverged' background expression conservation (EC) distribution.
}
\examples{

randomEC1 <- divergedEC(exprList$exprValues1, core_submatrix2_ordered,
                         ortho, EC$ECweights, conv = 0.001, maxIter = 200)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
