% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corM_fast.R
\name{get_corM}
\alias{get_corM}
\title{Construct a Gene-by-Gene Correlation Matrix}
\usage{
get_corM(expression, dropNArows = TRUE, threads = 8)
}
\arguments{
\item{expression}{A matrix of gene expression data. Rows represent genes and columns represent different conditions. The matrix should contain gene expression measurements across multiple conditions.}

\item{dropNArows}{Logical flag indicating whether to exclude rows where more than 50\% of the observations are missing. Defaults to \code{TRUE}. When set to \code{TRUE}, rows with more than 50\% missing values ('NA') are removed before computing the correlation matrix.}

\item{threads}{The number of threads to utilize for the correlation computation. This is particularly useful for handling large datasets. The correlation computation uses the \code{cor} function from the 'WGCNA' package.}
}
\value{
A square matrix representing the gene-by-gene correlations. The matrix is computed based on pairwise complete observations using Pearson's correlation method.
}
\description{
Generates a correlation matrix representing gene-gene relationships based on their expression across different conditions.
This function computes a 'gene X gene' correlation matrix from an input 'gene X condition' matrix.
}
\examples{
CorM1 <- get_corM(expression_matrix_1)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
