% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corM_fast.R
\name{get_corM_fast}
\alias{get_corM_fast}
\title{Construct Correlation Matrix (corM)}
\usage{
get_corM_fast(expression, dropNArows = TRUE, threads = 8)
}
\arguments{
\item{expression}{A matrix consisting of gene expression
measurements across multiple conditions (columns), for different genes (rows)}

\item{dropNArows}{option to indicate whether to remove rows of which more
than 50\% of the observations are missing. Defaults to "TRUE"}

\item{threads}{Number of threads to be used for the correlation calculation
using the 'cor' function in the 'WGCNA' package}
}
\value{
Square gene X gene correlation matrix.
}
\description{
This function makes a 'gene X gene' correlation matrix from a
'gene X condition' matrix
}
\examples{

CorM1 <- correlationMatrix(expression_matrix_1)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
