% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FECs.R
\name{get_expression_classes}
\alias{get_expression_classes}
\title{get functional expression classes}
\usage{
get_expression_classes(corM, coreGenes, height_cutoff = 1050)
}
\arguments{
\item{corM}{Correlation matrix}

\item{coreGenes}{Vector of core gene IDs}

\item{height_cutoff}{}
}
\description{
get functional expression classes
}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
