% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FECs.R
\name{get_expression_classes_ht}
\alias{get_expression_classes_ht}
\title{get functional expression classes}
\usage{
get_expression_classes_ht(corM, coreGenes, height_cutoff = 1050)
}
\arguments{
\item{corM}{Correlation matrix}

\item{coreGenes}{Vector of core gene IDs}

\item{height_cutoff}{}
}
\value{
heatmap
}
\description{
get functional expression classes
}
\examples{
heatmap <- get_expression_classes_ht(corM, coreGenes, height_cutoff = 1050)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
