% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEC.R
\name{getEC}
\alias{getEC}
\title{Calculate Expression Conservation (EC) Between Genes}
\usage{
getEC(m1, m2, conv = 0.001, maxIter = 200, threads = 1)
}
\arguments{
\item{m1}{A 'gene X gene' correlation matrix representing the first set of genes.}

\item{m2}{A 'gene X gene' correlation matrix representing the second set of genes. These are typically orthologous genes to those in \code{m1}.}

\item{conv}{Convergence criterion for the iterative process. The iteration stops when the change in EC value per gene between iterations is less than this threshold. Default value is 0.001.}

\item{maxIter}{Maximum number of iterations to perform. Default is 200.}

\item{threads}{Number of threads to use for computation. This parameter can significantly speed up the analysis on Linux and Mac computers. Default is set to 1 for compatibility with Windows systems.}
}
\value{
A list containing two elements:
\itemize{
\item \code{ECfinal}: A vector of the final EC scores for each gene, indicating the conservation of expression patterns.
\item \code{ECweights}: A vector of the final weights used in the last iteration of the weighted correlation calculation.
}
}
\description{
Computes the expression conservation (EC) between corresponding genes in two gene-gene correlation matrices. This function iteratively calculates a weighted correlation for each gene pair until convergence is reached.
}
\details{
The function first computes a simple correlation between corresponding genes in \code{m1} and \code{m2}. Iterative weighted correlations are then performed, updating the weights in each iteration based on the previous EC scores, until convergence is achieved based on the \code{conv} parameter.
}
\examples{
EC <- getEC(corM_ortho$csM1, csM2_ordered, conv = 0.001, maxIter = 200, threads = 1)
View(EC$ECfinal)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
