% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEC_fast.R
\name{getEC_fast}
\alias{getEC_fast}
\title{Get EC}
\usage{
getEC_fast(m1, m2, conv = 0.001, maxIter = 200, threads = 1)
}
\arguments{
\item{m1}{First 'gene X gene' correlation matrix}

\item{m2}{Second 'gene X gene' correlation matrix}

\item{conv}{Convergence criterion indicating how much the final EC value per gene can
differ from the result of the previous iteration. Default = 0.001}

\item{maxIter}{The maximum number of iterations. Default = 200}

\item{threads}{Number of threads to be used for WGCNA::cor and parallel:: mclapply.
This will heavily speed up the analysis on linux and MAC computers, but
defaults to 1 to support usage on windows computers.}
}
\value{
List (EClist) with the following elements:
\enumerate{
\item EClist$ECfinal = a vector of final converged EC scores per gene
\item ECfinal$ECweights = a vector of the last set of weights used for the iterative, weighted correlation
}
}
\description{
Calculate the expression conservation (EC) between (orthologuous) genes from two correlation matrices.
This function iteratively calculates the weighted correlation of a given row of m1
with a row (with the same rowindex) in m2 until convergence.
}
\examples{

EC <- getEC(corM_ortho$csM1, csM2_ordered, conv = 0.001, maxIter = 200, threads = 1)
View(ECfinal$ECfinal)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
