% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_subset_EC.R
\name{internal_remove_zero_variance_rows}
\alias{internal_remove_zero_variance_rows}
\title{Remove Rows and Corresponding Columns with Zero Variance}
\usage{
internal_remove_zero_variance_rows(mat1, mat2)
}
\arguments{
\item{mat1}{A numeric matrix for which rows and columns with zero variance are to be removed.}

\item{mat2}{A numeric matrix for which rows and columns with zero variance are to be removed.}
}
\value{
A named list with two elements, \code{matrixA} and \code{matrixB}, each being the input matrices
with rows and columns of zero variance removed.
}
\description{
This internal function identifies and removes rows (and their corresponding columns)
with zero variance from two matrices, ensuring the matrices remain symmetrical. It is
particularly useful for preprocessing correlation matrices before further analysis.
}
\examples{
mat1 <- matrix(rnorm(100), 10)
mat2 <- matrix(rnorm(100), 10)
result <- internal_remove_zero_variance_rows(mat1, mat2)
}
\keyword{internal}
