% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_EC_fast.R
\name{perfect_EC}
\alias{perfect_EC}
\title{Calculate 'Perfect' Expression Conservation (EC) Distribution}
\usage{
perfect_EC(
  exprM,
  labelsM,
  conv = 0.001,
  maxIter = 200,
  threads = 8,
  ortho,
  splits = 1,
  experiment_info,
  tolerance_thresh = 0.05
)
}
\arguments{
\item{exprM}{The first 'gene X condition' expression matrix.}

\item{labelsM}{A correlation (sub)matrix with gene IDs matching those in \code{exprM}.}

\item{conv}{Convergence threshold for the EC calculation process.}

\item{maxIter}{The maximum number of iterations for the EC calculations.}

\item{threads}{Number of threads to utilize for correlation calculations and row-wise EC calculations.}

\item{ortho}{Optional: A matrix containing IDs of genes in the first compendium (first column) and corresponding orthologs in the second compendium (second column).}

\item{splits}{The number of ways to split the dataset for EC calculation. Default is 1. When more than one, the function iteratively calculates EC over multiple splits of the dataset.}

\item{experiment_info}{Information about the experiments, used to guide the splitting of the dataset.}

\item{tolerance_thresh}{Tolerance threshold for balancing the size of each split.}
}
\value{
If \code{splits} is 1, a named vector of EC values representing perfect conservation. If \code{splits} is more than 1, a list containing EC values for each split, the mean EC value across splits, and variance of EC values per gene.
}
\description{
This function computes the 'perfect' expression conservation (EC) distribution by dividing a gene expression compendium into two parts and calculating the EC between the resulting matrices. The EC for each gene pair is obtained as a weighted correlation between corresponding rows in the two matrices.
}
\details{
The function initially splits the expression matrix \code{exprM} into two halves and computes the EC for these two parts. When \code{splits} is more than 1, the function performs multiple splits based on \code{experiment_info}, calculates EC for each split, and computes the mean and variance of EC values across splits.
}
\examples{
perfectEC1 <- perfect_EC(exprList$exprValues2, core_submatrix2_ordered,
conv = 0.001, maxIter = 200, threads = 8, ortho = singleCopyOrthologs_matrix)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
