% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_EC_fast.R
\name{perfect_EC_fast}
\alias{perfect_EC_fast}
\title{perfect EC distribution}
\usage{
perfect_EC_fast(
  exprM,
  labelsM,
  conv = 0.001,
  maxIter = 200,
  threads = 8,
  ortho,
  multiple_splits = TRUE,
  experiment_info,
  tolerance_thresh = 0.05
)
}
\arguments{
\item{exprM}{First 'gene X condition' expression matrix}

\item{labelsM}{Correlation (sub)matrix. Gene IDs must be the same as used in the exprM}

\item{conv}{Convergence treshold for EC calculations}

\item{maxIter}{Maximum number of iterations for EC calculations}

\item{threads}{Number of threads to use for cor calculations and row-wise EC calculations}

\item{ortho}{OPTIONAL argument. Matrix consisting of IDs of genes used in the
first compendium (first column) and gene IDs of the corresponding orthologs
found in the secondcompendium}
}
\value{
A  named vector of EC values representing perfect conservation
}
\description{
Calculate 'perfect' expression conservation (EC) distribution. This function splits a compendium
in two parts, and calculates the EC for the two resulting matrices (i.e the weighted correlation of
each row of matrix 1 (m1) with the corresponding row in matrix 2 (m2) is calculated).
}
\examples{

perfectEC1  <- perfect_EC(exprList$exprValues2, core_submatrix2_ordered, 
conv = 0.001, maxIter = 200, threads = 8, ortho = singleCopyOrthologs_matrix)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
