% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot distributions}
\usage{
plot_distributions(EC, conserved, diverged)
}
\arguments{
\item{EC}{Named vector of EC values}

\item{conserved}{Named vector of EC values representing conserved gene co-expression}

\item{diverged}{Named vector of EC values representing diverged gene co-expression}
}
\value{
A  plot of distributions
}
\description{
Plot distributions
}
\examples{

distr_plot <- plot_distributions(EC$ECfinal, perfectEC, randomEC_vec)
distr_plot

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
