% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read gene expression data}
\usage{
readData(dataset1, dataset2, NAstring = "NaN", source = "table")
}
\arguments{
\item{dataset1}{Tab-delimited file of log Fold Changes for different genes (rows) measured under different conditions (columns)}

\item{dataset2}{Tab-delimited file of log Fold Changes for different genes (rows) measured under different conditions (columns)}

\item{NAstring}{The NA string used (e.g "NaN")}

\item{source}{The data source of which the expression compendium was retrieved (e.g COLOMBOS).
Default: expression table (tab delimited file of gene expression measurements) containing row IDs in the first column, and experiment
IDs in the first row}
}
\value{
A list (exprList) of 6 named items:
\enumerate{
\item $baseName1, $baseName2: basename of the file
\item $exprValues1, $exprValues2: numeric dataframe or matrix of expression values
\item $headerInfo1, $headerInfo2: dataframe containing additional headerlines explaning the experimental conditions
}
}
\description{
This function can be used to read a gene expression compendium (a large table of log Fold Changes)
and store it in an R dataframe. If the data is retrieved from COLOMBOS, the option source = "COLOMBOS"
can be used. By default (for other data sources), a file tab delimited file must be supplied consisting of
gene expression measurements for different genes (rows), across multiple conditions (columns). The first
row must consist of the condition identifiers, while the first column must contain gene IDs.
}
\examples{
# read data

exprList <- readData(dataset1, dataset2, rownamePos1 = 1,
rownamePos2 = 1, NAstring = "NaN", source = "table")

# inspect the dataframe of expression values

View(head(exprList$exprValues1))

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
