% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{singleCopyOrthologs}
\alias{singleCopyOrthologs}
\title{Orthology Information for Salmonella Typhimurium}
\format{
A data frame with two columns representing orthologous gene pairs.
}
\usage{
singleCopyOrthologs
}
\description{
This dataset provides orthology information between gene pairs.
It consists of two columns, each representing a gene ID, where every pair of gene IDs corresponds to an ortholog pair.
The data can be used to compare gene expression between different strains or species.
}
\keyword{datasets}
