% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_matrix.R
\name{sort_matrix}
\alias{sort_matrix}
\title{Sort matrices}
\usage{
sort_matrix(csM1, csM2, ortho)
}
\arguments{
\item{ortho}{OPTIONAL argument. Two-column dataframe of gene IDs in species1/sample1,
and the corresponding orthologous gene IDs in species2/sample2.}

\item{matrix1}{First 'gene X gene' correlation matrix}

\item{matrix2}{Second 'gene X gene' correlation matrix}
}
\value{
A matrix (csM2_ordered) that is row- and column ordered according to matrix1
(either by identical gene IDs or ortholog gene IDs if the optional argument 'ortho' is provided)
}
\description{
This function sorts rows and columns of a symmetrical correlation matrix
(second argument; csM2) according to the other (first argument; csM1).
As a result, row n in Matrix 1 and row n in Matrix 2 correspond to the
same gene or ortholog, with the row values describing the correlation
of this gene with the other genes in the compendium in the exact same order.
}
\examples{

# example when matrix 2 (csM2) has to be ordered according to matrix 1 (csM1)
# the orthology info (corM_ortho$orthologs) is taken into account

csM2_ordered <- sort_matrix(corM_ortho$csM1, corM_ortho$csM2, corM_ortho$orthologs)

# example when matrix 2 (csM2) has to be ordered according to matrix 1 (csM1)
# bot matrices contain identical gene IDs

csM2_ordered <- sort_matrix(corM_ortho$csM1, corM_ortho$csM2)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
