% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sTyph_LT2}
\alias{sTyph_LT2}
\title{Gene Expression Data from Salmonella Typhimurium Strain LT2}
\format{
A data frame with 200 rows (genes) and 100 columns (expression values).
}
\usage{
sTyph_LT2
}
\description{
This dataset contains gene expression data for Salmonella Typhimurium strain LT2.
A subset of 200 genes and 100 columns was selected for inclusion in this dataset.
The data was sourced from http://colombos.net/.
}
\keyword{datasets}
