% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGO_salmonella.R
\name{topGO_salmonella}
\alias{topGO_salmonella}
\title{perfect EC distribution}
\usage{
topGO_salmonella(
  goMappingPath = "examples/typhimurium_geneid2go.map",
  output_path,
  filename = "go_output",
  myInterestingGenes,
  ontology_type = "BP",
  top_nodes = 10
)
}
\arguments{
\item{goMappingPath}{Path to the GO mapping file
(i.e. a file that associates each gene id with a set of GO terms)}

\item{myInterestingGenes}{Vector of gene IDs which represent the subset of genes of interest}

\item{ontology_type}{String indicating which of the three ontologies should be used.
(Options: "BP", "MF" and "CC", which refers to Biological Process, Molecular Function and Cellular Component)}

\item{top_nodes}{Number of top nodes that should be shown (Shows the 10 top scoring terms by default)}

\item{outputPath}{Path for writing output files}
}
\value{
A  named vector of EC values representing perfect conservation
}
\description{
Wrapper function to perform GO enrichment for Salmonella genes using the TopGO package
}
\examples{

go_enrichment <- topGO_salmonella(goMappingPath, output_path,
filename = "EC_diverged_GO",myInterestingGenes)

}
\author{
Wim Cuypers, \email{wim.cuypers@uantwerpen.be}
}
